changed location types from i32 to f64 added blockedcomponent

+++ src/builder/dungeon.rs
@@ -14,7 +14,7 @@ pub fn tear_down_level(ecs_: &mut ecs::ECS) {
     }
 }
 
-pub fn create_floor_tile(ecs_: &mut ecs::ECS, x: i32, y: i32) -> ecs::Entity {
+pub fn create_floor_tile(ecs_: &mut ecs::ECS, x: f64, y: f64) -> ecs::Entity {
     let entity = ecs_.allocator.allocate();
     ecs_.dungeon_component.set(entity, components::DungeonComponent {
         type_: dungeon::DungeonElement::Floor
@@ -27,7 +27,7 @@ pub fn create_floor_tile(ecs_: &mut ecs::ECS, x: i32, y: i32) -> ecs::Entity {
     });
     ecs_.render_component.set(entity, components::RenderComponent {
         base_sprite: "default",
-        base_sprite_size: 1,
+        base_sprite_size: 1.0,
         animation: None,
         visible: true,
         render_layer: 0
@@ -35,7 +35,7 @@ pub fn create_floor_tile(ecs_: &mut ecs::ECS, x: i32, y: i32) -> ecs::Entity {
     entity
 }
 
-pub fn create_wall_tile(ecs_: &mut ecs::ECS, x: i32, y: i32) -> ecs::Entity {
+pub fn create_wall_tile(ecs_: &mut ecs::ECS, x: f64, y: f64) -> ecs::Entity {
     let entity = ecs_.allocator.allocate();
     ecs_.dungeon_component.set(entity, components::DungeonComponent {
         type_: dungeon::DungeonElement::Wall
@@ -49,7 +49,7 @@ pub fn create_wall_tile(ecs_: &mut ecs::ECS, x: i32, y: i32) -> ecs::Entity {
     entity
 }
 
-pub fn create_connector_tile(ecs_: &mut ecs::ECS, x: i32, y: i32) -> ecs::Entity {
+pub fn create_connector_tile(ecs_: &mut ecs::ECS, x: f64, y: f64) -> ecs::Entity {
     let entity = ecs_.allocator.allocate();
     ecs_.dungeon_component.set(entity, components::DungeonComponent {
         type_: dungeon::DungeonElement::Connector
